Move AnnotatedPncUpdateDatasetParsedXml type into root types folder
